requirements.txt
pydantic-ai
pandas
python-dotenv
google-generativeai